requirements.txt
fastapi[standard]
pymupdf
Pillow
pytesseract
requests